requirements.txt
click==7.1.2
fastapi==0.64.0
h11==0.12.0
pydantic==1.8.1
starlette==0.13.6
typing-extensions==3.10.0.0
uvicorn==0.13.4
firebase_admin
pytest
pytest-asyncio
gunicorn
google-cloud-logging